requirements.txt
setuptools==57.1.0
ale-py==0.7.1
einops==0.3.2
gym[atari, accept-rom-license]==0.21.0
hydra-core==1.1.1
opencv-python
protobuf==3.20.*
psutil==5.8.0
pygame==2.1.2
tqdm==4.62.3
wandb==0.12.7
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
